+++ pkg/query/columnquery.go
@@ -67,11 +67,11 @@ func NewColumnQueryAPI(
 		tracer:             tracer,
 		shareClient:        shareClient,
 		querier:            querier,
-		tableConverterPool: newTableConverterPool(),
+		tableConverterPool: NewTableConverterPool(),
 	}
 }
 
-func newTableConverterPool() *sync.Pool {
+func NewTableConverterPool() *sync.Pool {
 	return &sync.Pool{
 		New: func() any {
 			return &tableConverter{
@@ -302,14 +302,14 @@ func RenderReport(
 			Report:   &pb.QueryResponse_Pprof{Pprof: buf.Bytes()},
 		}, nil
 	case pb.QueryRequest_REPORT_TYPE_TOP:
-		top, err := GenerateTopTable(ctx, p)
+		top, cumulative, err := GenerateTopTable(ctx, p)
 		if err != nil {
 			return nil, status.Errorf(codes.Internal, "failed to generate pprof: %v", err.Error())
 		}
 
 		return &pb.QueryResponse{
 			//nolint:staticcheck // SA1019: TODO: The cumulative should be passed differently in the future.
-			Total:    int64(top.Total),
+			Total:    cumulative,
 			Filtered: filtered,
 			Report:   &pb.QueryResponse_Top{Top: top},
 		}, nil