+++ ui/packages/shared/components/src/Dropdown/index.tsx
@@ -25,17 +25,17 @@ const Dropdown = ({
   variant = 'neutral',
   children,
 }: {
-  text: string;
+  text?: string;
   element?: React.ReactNode;
   variant?: 'neutral' | 'link' | 'primary' | undefined;
   children?: React.ReactNode;
 }): JSX.Element => {
   return (
     <>
-      <Menu as="div" className="relative inline-block text-left">
+      <Menu as="div" className="relative z-[999999999] inline-block text-left">
         <div>
           <Menu.Button as="div">
-            {element !== undefined ? (
+            {element === undefined ? (
               <Button variant={variant}>
                 {text}{' '}
                 <Icon
@@ -58,7 +58,7 @@ const Dropdown = ({
           leaveFrom="transform opacity-100 scale-100"
           leaveTo="transform opacity-0 scale-95"
         >
-          <Menu.Items className="absolute right-0 mt-2 w-56 origin-top-right divide-y divide-gray-100 rounded-md bg-white shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none">
+          <Menu.Items className="absolute right-0 mt-2 w-36 origin-top-right divide-y divide-gray-100 rounded-md bg-white shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none dark:bg-gray-900">
             <div className="px-1 py-1 ">{children}</div>
           </Menu.Items>
         </Transition>
@@ -70,17 +70,20 @@ const Dropdown = ({
 const Item = ({
   children,
   onSelect,
+  selected,
 }: {
   children: React.ReactNode;
   onSelect: () => void;
+  selected?: boolean;
 }): JSX.Element => {
   return (
     <Menu.Item>
       {({active}) => (
         <button
           className={cx(
-            active ? 'bg-indigo-500 text-white' : 'text-gray-900',
-            'group flex w-full items-center rounded-md px-2 py-2 text-sm'
+            active ? 'bg-indigo-500 text-white' : 'text-gray-900 dark:text-white',
+            'group mb-px flex w-full items-center rounded-md px-2 py-2 text-sm',
+            selected != null && selected ? 'bg-indigo-500 font-bold !text-white' : ''
           )}
           onClick={onSelect}
         >