+++ ui/packages/app/grafana-datasource-plugin/CHANGELOG.md
@@ -3,6 +3,98 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.136](https://github.com/parca-dev/parca/compare/@parca/grafana-datasource@0.0.135...@parca/grafana-datasource@0.0.136) (2023-06-14)
+
+**Note:** Version bump only for package @parca/grafana-datasource
+
+## [0.0.135](https://github.com/parca-dev/parca/compare/@parca/grafana-datasource@0.0.134...@parca/grafana-datasource@0.0.135) (2023-06-13)
+
+**Note:** Version bump only for package @parca/grafana-datasource
+
+## [0.0.134](https://github.com/parca-dev/parca/compare/@parca/grafana-datasource@0.0.133...@parca/grafana-datasource@0.0.134) (2023-06-12)
+
+**Note:** Version bump only for package @parca/grafana-datasource
+
+## 0.0.133 (2023-06-10)
+
+**Note:** Version bump only for package @parca/grafana-datasource
+
+## [0.0.132](https://github.com/parca-dev/parca/compare/@parca/grafana-datasource@0.0.131...@parca/grafana-datasource@0.0.132) (2023-06-07)
+
+**Note:** Version bump only for package @parca/grafana-datasource
+
+## [0.0.131](https://github.com/parca-dev/parca/compare/@parca/grafana-datasource@0.0.130...@parca/grafana-datasource@0.0.131) (2023-06-07)
+
+**Note:** Version bump only for package @parca/grafana-datasource
+
+## 0.0.130 (2023-06-06)
+
+**Note:** Version bump only for package @parca/grafana-datasource
+
+## 0.0.129 (2023-06-05)
+
+**Note:** Version bump only for package @parca/grafana-datasource
+
+## [0.0.128](https://github.com/parca-dev/parca/compare/@parca/grafana-datasource@0.0.127...@parca/grafana-datasource@0.0.128) (2023-06-01)
+
+**Note:** Version bump only for package @parca/grafana-datasource
+
+## [0.0.127](https://github.com/parca-dev/parca/compare/@parca/grafana-datasource@0.0.126...@parca/grafana-datasource@0.0.127) (2023-06-01)
+
+**Note:** Version bump only for package @parca/grafana-datasource
+
+## 0.0.126 (2023-06-01)
+
+**Note:** Version bump only for package @parca/grafana-datasource
+
+## 0.0.125 (2023-05-30)
+
+**Note:** Version bump only for package @parca/grafana-datasource
+
+## 0.0.124 (2023-05-26)
+
+**Note:** Version bump only for package @parca/grafana-datasource
+
+## 0.0.123 (2023-05-25)
+
+**Note:** Version bump only for package @parca/grafana-datasource
+
+## 0.0.122 (2023-05-10)
+
+**Note:** Version bump only for package @parca/grafana-datasource
+
+## [0.0.121](https://github.com/parca-dev/parca/compare/@parca/grafana-datasource@0.0.120...@parca/grafana-datasource@0.0.121) (2023-05-04)
+
+**Note:** Version bump only for package @parca/grafana-datasource
+
+## 0.0.120 (2023-05-04)
+
+**Note:** Version bump only for package @parca/grafana-datasource
+
+## 0.0.119 (2023-05-03)
+
+**Note:** Version bump only for package @parca/grafana-datasource
+
+## [0.0.118](https://github.com/parca-dev/parca/compare/@parca/grafana-datasource@0.0.117...@parca/grafana-datasource@0.0.118) (2023-04-18)
+
+**Note:** Version bump only for package @parca/grafana-datasource
+
+## 0.0.117 (2023-04-18)
+
+**Note:** Version bump only for package @parca/grafana-datasource
+
+## 0.0.116 (2023-04-08)
+
+**Note:** Version bump only for package @parca/grafana-datasource
+
+## [0.0.115](https://github.com/parca-dev/parca/compare/@parca/grafana-datasource@0.0.114...@parca/grafana-datasource@0.0.115) (2023-04-06)
+
+**Note:** Version bump only for package @parca/grafana-datasource
+
+## 0.0.114 (2023-04-06)
+
+**Note:** Version bump only for package @parca/grafana-datasource
+
 ## 0.0.113 (2023-04-04)
 
 **Note:** Version bump only for package @parca/grafana-datasource