+++ ui/packages/shared/hooks/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.0.8 (2023-05-03)
+
+**Note:** Version bump only for package @parca/hooks
+
+## 0.0.7 (2023-04-18)
+
+**Note:** Version bump only for package @parca/hooks
+
 ## 0.0.6 (2023-03-29)
 
 **Note:** Version bump only for package @parca/hooks